+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Forms integration with Aptean Respond API
+
 ## 8.0 (2020-03-30)
 
 Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/6.0...7.0